Organize the e2e test files, and also create the cases for the US8 filters.
