+++ server/test/e2e_US4.test.ts
@@ -3,8 +3,9 @@ import { StatusCodes } from "http-status-codes";
 import request from "supertest";
 import app from "../src/app";
 import { Database } from "../src/database";
-import { DocumentType, Stakeholder } from "../src/model/document";
+import { DocumentType } from "../src/model/document";
 import { ScaleType } from "../src/model/scale";
+import { Stakeholder } from "../src/model/stakeholder";
 import { loginAsPlanner, loginAsResident } from "./utils";
 dotenv.config();
